@@ -4,13 +4,11 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:image_picker/image_picker.dart';
 import 'package:permission_handler/permission_handler.dart';
-import 'package:r_scan_example/scan_dialog.dart';
-import 'scan_camera_dialog.dart';
 import 'package:r_scan/r_scan.dart';
 
-void main() async {
-  WidgetsFlutterBinding.ensureInitialized();
+import 'scan_camera_dialog.dart';
 
+void main() {
   runApp(MyApp());
 }
 
@@ -22,9 +20,7 @@ class MyApp extends StatefulWidget {
 class _MyAppState extends State<MyApp> {
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(
-      home: MyPage(),
-    );
+    return MaterialApp(home: MyPage());
   }
 }
 
@@ -34,89 +30,88 @@ class MyPage extends StatefulWidget {
 }
 
 class _MyPageState extends State<MyPage> {
-  RScanResult result;
+  final ImagePicker picker = ImagePicker();
+
+  RScanResult? result;
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(
-        title: Text('scan example'),
-      ),
+      appBar: AppBar(title: const Text('scan example')),
       body: Column(
         mainAxisAlignment: MainAxisAlignment.center,
         crossAxisAlignment: CrossAxisAlignment.center,
         children: <Widget>[
           Center(
-              child: Text(result == null
+            child: Text(
+              result == null
                   ? '点击下方按钮开始扫码'
-                  : '扫码结果${result.toString().split(',').join('\n')}')),
+                  : '扫码结果${result.toString().split(',').join('\n')}',
+            ),
+          ),
           Center(
-            child: FlatButton(
+            child: TextButton(
               onPressed: () async {
-                final result = await Navigator.of(context).push(
-                    MaterialPageRoute(
-                        builder: (BuildContext context) =>
-                            RScanCameraDialog()));
+                final RScanResult? result = await Navigator.of(context).push(
+                  MaterialPageRoute<RScanResult>(
+                    builder: (BuildContext context) => RScanCameraDialog(),
+                  ),
+                );
                 setState(() {
                   this.result = result;
                 });
               },
-              child: Text('RScanCamera开始扫码'),
+              child: const Text('RScanCamera 开始扫码'),
             ),
           ),
           Center(
-            child: FlatButton(
-              onPressed: () async {
-                final result = await Navigator.of(context).push(
-                    MaterialPageRoute(
-                        builder: (BuildContext context) => RScanDialog()));
-                setState(() {
-                  this.result = result;
-                });
-              },
-              child: Text('RScanView开始扫码(已弃用)'),
-            ),
-          ),
-          Center(
-            child: FlatButton(
+            child: TextButton(
               onPressed: () async {
                 if (await canReadStorage()) {
-                  var image =
-                      await ImagePicker.pickImage(source: ImageSource.gallery);
+                  final PickedFile? image = await picker.getImage(
+                    source: ImageSource.gallery,
+                  );
                   if (image != null) {
-                    final result = await RScan.scanImagePath(image.path);
+                    final RScanResult? result = await RScan.scanImagePath(
+                      image.path,
+                    );
                     setState(() {
                       this.result = result;
                     });
                   }
                 }
               },
-              child: Text('选择图片扫描'),
+              child: const Text('选择图片扫描'),
             ),
           ),
           Center(
-            child: FlatButton(
+            child: TextButton(
               onPressed: () async {
-                final result = await RScan.scanImageUrl(
-                    "https://s.cn.bing.net/th?id=OJ.5F0gxqWmxskS0Q&w=75&h=75&pid=MSNJVFeeds");
+                final RScanResult? result = await RScan.scanImageUrl(
+                  'https://s.cn.bing.net/th'
+                  '?id=OJ.5F0gxqWmxskS0Q&w=75&h=75&pid=MSNJVFeeds',
+                );
                 setState(() {
                   this.result = result;
                 });
               },
-              child: Text('网络图片解析'),
+              child: const Text('网络图片解析'),
             ),
           ),
           Center(
-            child: FlatButton(
+            child: TextButton(
               onPressed: () async {
-                ByteData data = await rootBundle.load('images/qrCode.png');
-                final result =
-                    await RScan.scanImageMemory(data.buffer.asUint8List());
+                final ByteData data = await rootBundle.load(
+                  'images/qrCode.png',
+                );
+                final RScanResult? result = await RScan.scanImageMemory(
+                  data.buffer.asUint8List(),
+                );
                 setState(() {
                   this.result = result;
                 });
               },
-              child: Text('内存图片解析'),
+              child: const Text('内存图片解析'),
             ),
           ),
         ],
@@ -125,20 +120,12 @@ class _MyPageState extends State<MyPage> {
   }
 
   Future<bool> canReadStorage() async {
-    if (Platform.isIOS) return true;
-    var status = await PermissionHandler()
-        .checkPermissionStatus(PermissionGroup.storage);
-    if (status != PermissionStatus.granted) {
-      var future = await PermissionHandler()
-          .requestPermissions([PermissionGroup.storage]);
-      for (final item in future.entries) {
-        if (item.value != PermissionStatus.granted) {
-          return false;
-        }
-      }
-    } else {
+    if (Platform.isIOS) {
       return true;
     }
-    return true;
+    final Map<Permission, PermissionStatus> status = await <Permission>[
+      Permission.storage,
+    ].request();
+    return status[Permission.storage] == PermissionStatus.granted;
   }
 }
