@@ -1,9 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:permission_handler/permission_handler.dart';
-import 'package:r_scan_example/scan_dialog.dart';
-import 'package:r_scan/src/r_scan_camera.dart';
+import 'package:r_scan/r_scan.dart';
 
-List<RScanCameraDescription> rScanCameras;
+import 'scan_dialog.dart';
+
+List<RScanCameraDescription>? rScanCameras;
 
 class RScanCameraDialog extends StatefulWidget {
   @override
@@ -11,31 +12,30 @@ class RScanCameraDialog extends StatefulWidget {
 }
 
 class _RScanCameraDialogState extends State<RScanCameraDialog> {
-  RScanCameraController _controller;
+  RScanCameraController? _controller;
   bool isFirst = true;
 
-  void initCamera() async {
-    if (rScanCameras == null || rScanCameras.length == 0) {
-      final result = await PermissionHandler()
-          .checkPermissionStatus(PermissionGroup.camera);
-      if (result == PermissionStatus.granted) {
+  Future<void> initCamera() async {
+    if (rScanCameras?.isEmpty != false) {
+      final Map<Permission, PermissionStatus> status = await <Permission>[
+        Permission.camera,
+      ].request();
+      if (status[Permission.camera] == PermissionStatus.granted) {
         rScanCameras = await availableRScanCameras();
-        print('返回可用的相机：${rScanCameras.join('\n')}');
+        print('返回可用的相机：${rScanCameras!.join('\n')}');
       } else {
-        final resultMap = await PermissionHandler()
-            .requestPermissions([PermissionGroup.camera]);
-        if (resultMap[PermissionGroup.camera] == PermissionStatus.granted) {
-          rScanCameras = await availableRScanCameras();
-        } else {
-          print('相机权限被拒绝，无法使用');
-        }
+        print('相机权限被拒绝，无法使用');
+        return;
       }
     }
-    if (rScanCameras != null && rScanCameras.length > 0) {
+
+    if (rScanCameras?.isNotEmpty == true) {
       _controller = RScanCameraController(
-          rScanCameras[0], RScanCameraResolutionPreset.high)
+        rScanCameras![0],
+        RScanCameraResolutionPreset.high,
+      )
         ..addListener(() {
-          final result = _controller.result;
+          final RScanResult? result = _controller!.result;
           if (result != null) {
             if (isFirst) {
               Navigator.of(context).pop(result);
@@ -44,10 +44,9 @@ class _RScanCameraDialogState extends State<RScanCameraDialog> {
           }
         })
         ..initialize().then((_) {
-          if (!mounted) {
-            return;
+          if (mounted) {
+            setState(() {});
           }
-          setState(() {});
         });
     }
   }
@@ -66,16 +65,16 @@ class _RScanCameraDialogState extends State<RScanCameraDialog> {
 
   @override
   Widget build(BuildContext context) {
-    if (rScanCameras == null || rScanCameras.length == 0) {
+    if (rScanCameras?.isEmpty != false) {
       return Scaffold(
         body: Container(
           alignment: Alignment.center,
-          child: Text('not have available camera'),
+          child: const Text('not have available camera'),
         ),
       );
     }
-    if (!_controller.value.isInitialized) {
-      return Container();
+    if (_controller?.value.isInitialized != true) {
+      return const SizedBox.shrink();
     }
     return Scaffold(
       backgroundColor: Colors.black,
@@ -83,16 +82,17 @@ class _RScanCameraDialogState extends State<RScanCameraDialog> {
         children: <Widget>[
           ScanImageView(
             child: AspectRatio(
-              aspectRatio: _controller.value.aspectRatio,
-              child: RScanCamera(_controller),
+              aspectRatio: 1 / _controller!.value.aspectRatio,
+              child: RScanCamera(_controller!),
             ),
           ),
           Align(
-              alignment: Alignment.bottomCenter,
-              child: FutureBuilder(
-                future: getFlashMode(),
-                builder: _buildFlashBtn,
-              ))
+            alignment: Alignment.bottomCenter,
+            child: FutureBuilder<bool>(
+              future: getFlashMode(),
+              builder: _buildFlashBtn,
+            ),
+          )
         ],
       ),
     );
@@ -101,29 +101,32 @@ class _RScanCameraDialogState extends State<RScanCameraDialog> {
   Future<bool> getFlashMode() async {
     bool isOpen = false;
     try {
-      isOpen = await _controller.getFlashMode();
+      isOpen = (await _controller?.getFlashMode()) ?? false;
     } catch (_) {}
     return isOpen;
   }
 
   Widget _buildFlashBtn(BuildContext context, AsyncSnapshot<bool> snapshot) {
-    return snapshot.hasData
-        ? Padding(
-            padding: EdgeInsets.only(
-                bottom: 24 + MediaQuery.of(context).padding.bottom),
-            child: IconButton(
-                icon: Icon(snapshot.data ? Icons.flash_on : Icons.flash_off),
-                color: Colors.white,
-                iconSize: 46,
-                onPressed: () {
-                  if (snapshot.data) {
-                    _controller.setFlashMode(false);
-                  } else {
-                    _controller.setFlashMode(true);
-                  }
-                  setState(() {});
-                }),
-          )
-        : Container();
+    if (!snapshot.hasData) {
+      return const SizedBox.shrink();
+    }
+    return Padding(
+      padding: EdgeInsets.only(
+        bottom: 24 + MediaQuery.of(context).padding.bottom,
+      ),
+      child: IconButton(
+        icon: Icon(snapshot.data! ? Icons.flash_on : Icons.flash_off),
+        color: Colors.white,
+        iconSize: 46,
+        onPressed: () {
+          if (snapshot.data!) {
+            _controller?.setFlashMode(false);
+          } else {
+            _controller?.setFlashMode(true);
+          }
+          setState(() {});
+        },
+      ),
+    );
   }
 }
